Added filtering options and search bar
The dropdowns are a little broken because the list of items do not cover the second component below them which needs to be fixed. I also don't like the styling of the buttons. I want the chevron icon to be in the right hand corner but there's so much CSS in my mind right now I want to die...