Added photo and database pulling functionality
Fixed the database and changed the fetchWines function to its own hook that gets called all the damn time. Still need to do some testing to see if it pulls new data when a wine is created on another client and a user on their client swaps between cellar and ratings views.